Got metal working (and not working) for 0.0.5
