+++ Pika/Views/Splash/MetalShader.metal
@@ -3,9 +3,6 @@
 #include <metal_stdlib>
 #include <simd/simd.h>
 
-// Including header shared between this Metal shader code and Swift/C code executing Metal API commands
-//#import "ShaderTypes.h"
-
 using namespace metal;
 
 // macros
